+++ types/open_ai_chat.go
@@ -271,3 +271,110 @@ func (c *OpenAIChat) UpdateSystem(systemPrompt string) {
 func (c *OpenAIChat) UpdateTemperature(newValue float32) {
 	c.Temperature = newValue
 }
+
+func (c *OpenAIChat) WithJsonSchema(message string, schemaName string, schema map[string]interface{}, onUpdate ChatAIMessageChunkReceiver) error {
+	apiKey := strings.TrimSpace(c.ApiKey)
+	if apiKey == "" {
+		return fmt.Errorf("no OpenAI api key defined")
+	}
+
+	model := strings.TrimSpace(strings.ToLower(c.Model))
+	if model == "" {
+		return fmt.Errorf("no chat ai model defined")
+	}
+
+	messages := []OpenAIChatMessage{}
+
+	if c.SystemPrompt != "" {
+		systemMessage := OpenAIChatMessage{
+			Content: c.SystemPrompt,
+			Role:    "system",
+		}
+
+		messages = append(messages, systemMessage)
+	}
+
+	userMessage := OpenAIChatMessage{
+		Content: message,
+		Role:    "user",
+	}
+
+	messages = append(messages, c.Conversation...)
+	messages = append(messages, userMessage)
+
+	url := "https://api.openai.com/v1/chat/completions"
+
+	body := map[string]interface{}{
+		"model":    model,
+		"messages": messages,
+		"stream":   false,
+		"response_format": map[string]interface{}{
+			"type": "json_schema",
+			"json_schema": map[string]interface{}{
+				"name":   schemaName,
+				"schema": schema,
+			},
+		},
+		"temperature": c.Temperature,
+	}
+
+	jsonData, err := json.Marshal(&body)
+	if err != nil {
+		return err
+	}
+
+	req, err := http.NewRequest("POST", url, bytes.NewBuffer([]byte(jsonData)))
+	if err != nil {
+		return err
+	}
+
+	// setup ...
+	req.Header.Set("Content-Type", "application/json")
+	req.Header.Set("Authorization", fmt.Sprintf("Bearer %s", apiKey))
+	// ... and finally send the JSON data
+	client := &http.Client{}
+	resp, err := client.Do(req)
+	if err != nil {
+		return err
+	}
+	defer resp.Body.Close()
+
+	if resp.StatusCode != 200 {
+		return fmt.Errorf("unexpected response %v", resp.StatusCode)
+	}
+
+	// load the response
+	responseData, err := io.ReadAll(resp.Body)
+	if err != nil {
+		return err
+	}
+
+	var chatResponse OpenAIChatCompletionResponseV1
+	err = json.Unmarshal(responseData, &chatResponse)
+	if err != nil {
+		return err
+	}
+
+	assistantMessage := OpenAIChatMessage{
+		Content: "",
+		Role:    "assistant",
+	}
+	if len(chatResponse.Choices) > 0 {
+		assistantMessage.Content = chatResponse.Choices[0].Message.Content
+		assistantMessage.Role = chatResponse.Choices[0].Message.Role
+	}
+
+	c.Conversation = append(
+		c.Conversation,
+		userMessage, assistantMessage,
+	)
+
+	err = onUpdate(assistantMessage.Content)
+	if err != nil {
+		return err
+	}
+
+	c.TotalTokens += chatResponse.Usage.TotalTokens
+
+	return nil
+}