+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log (go-package-manager)
 
+## 0.22.0
+
+- **BREAKING CHANGE**: default AI model for Open AI is `gpt-4o-mini` now
+- **BREAKING CHANGE**: default AI model for Ollama is `llama3.3` now
+- feat: `generate project` command that generates a Go project with natural language
+
 ## 0.21.3
 
 - ci: automatic update of [GitHub wiki](https://github.com/mkloubert/go-package-manager/wiki) as well